refactor(shell): extract sidebar account controls
